Add deploy input option
This new option defaults to true, which means it will always deploy if necessary. If set to false, the user can explicitly disable deploy for the deploy-action which is useful in cases such as when using a hibernating deployment as a template the user would first have to wake up before deploying.